+++ tests/e2e/modules/clock_spec.js
@@ -119,6 +119,14 @@ describe("Clock module", function () {
 			return app.client.waitUntilWindowLoaded()
 				.getText(".clock .week").should.eventually.match(weekRegex);
 		});
+
+		it("shows week with correct number of week of year", function() {
+			const currentWeekNumber = require("current-week-number")();
+			const weekToShow = "Week " + currentWeekNumber;
+			return app.client.waitUntilWindowLoaded()
+				.getText(".clock .week").should.eventually.equal(weekToShow);
+		});
+
 	});
 
 });